+++ pages/dev_unlocalized/src/payload.config.ts
@@ -49,7 +49,7 @@ export default buildConfig({
         path && process.env.NEXT_PUBLIC_FRONTEND_URL
           ? `${process.env.NEXT_PUBLIC_FRONTEND_URL}${preview ? '/preview' : ''}${path}`
           : null,
-      // Disable the KV cache for findPageByPath/resolvePagePath lookups
+      // Disable the KV cache for findPageByPath lookups
       pathCache: false,
     }),
   ],